refactor(app): mod.rs had five lines of headroom; extract instead of bump
**F18** of the pre-2.1 review (`F-tests-guards.md`) — filed `informational` at
1448 of 1500. It is **1495 of 1500 on HEAD**: five lines.

That is the state the guard exists to prevent. Its own message says "Extract
logic into a src/app/ child module instead of growing mod.rs. Don't just raise
CEILING" — but the next person to add a `Message` variant meets that message
mid-feature, under deadline, one line from green, with a constant right there.
The ratchet history shows what the number is supposed to feel like: 4000 → 1500
was set when the sweep left mod.rs at ~1076, i.e. with ~420 lines of slack.

## Two verbatim relocations, both into a module that already claimed them

**The pure predicates → `util.rs` (95 lines).** `place_pty_cursor_from_screen`,
`POST_CHORD_BOUNCE_WINDOW`, `is_post_chord_bounce`, `should_fire_refresh`,
`is_spyc_meta_when_pane_focused`. None touches `App`, `Runtime` or `ViewState` —
they take plain arguments and return a decision. `util.rs`'s own header already
drew this line: *"Leaf helper functions with no `App`/`Runtime` dependency … the
app-domain glue (`sh_c` → `Effect`, `row_from_entry` → `RowData`) stays in
mod.rs."* These were on the wrong side of it. `sh_c` and `row_from_entry` stay,
as that sentence says.

**The loop-scoped types → `run.rs` (125 lines).** `Draw` + `impl Draw`,
`DispatchFlow`, `RunCtx` + its `#[cfg(test)] for_test`. Their own docs say where
they belong: `RunCtx` is "Run()-scoped scratch for the event loop … built by
`App::run_setup`, dropped when `run()` returns"; `DispatchFlow` is "the result of
`App::dispatch_effective`". `run.rs` owns that loop.

`RunCtx`'s fields are read by the sibling loop-step modules (`loop_steps`,
`sources`, `streaming`, `session`, …), so they are now `pub(super)` — from
`app::run` that is **exactly** the reachability they had as private items of
`app`: the whole `app` subtree. Same visibility, expressed one module deeper, not
a widening. mod.rs re-exports the type (`pub use run::RunCtx;`) so no call site
had to learn it moved.

**1495 → 1278.** Ceiling ratcheted 1500 → 1350.

## Why the headroom stays small on purpose

52% of what remains is `///` field docs on `Message`, `Runtime` and `ViewState` —
which is what AGENTS.md says this module is *for* ("the three struct defs +
`Message` enum + small glue"). Only 23 lines are plain `//` comments. So 1278 is
close to the floor for the current design, and 1350 leaves room for a new
`Message` variant with its documentation but not for a feature's logic. That is
the pressure the guard is supposed to apply.

Stripping those docs to buy room would be the wrong trade twice over: they are
rustdoc attached to the field they explain, and this very review kept finding
what happens when a comment drifts from its subject (F11's doc block spliced onto
the wrong guard, #400's four copies of one claim with one updated).

## Verification

- **Behaviour-preserving:** no test assertion changed in the diff — the
  relocation rule from AGENTS.md.
- **The ceiling bites:** padding mod.rs to 1358 fails with `src/app/mod.rs is
  1358 lines, over the 1350-line anti-monolith ceiling`. Reverted.
- One incidental fix: `KeyEvent` reached the harness fixtures through
  `use super::*` because mod.rs imported it for the predicates. Those fixtures
  now name it themselves, rather than keeping a production-unused import alive
  upstream to feed them.

`make check-ci` → `=== GATE: PASS ===`.
